"playable" pong

+++ core/src/com/gdx/jpong/model/GameObject.java
@@ -3,10 +3,10 @@ package com.gdx.jpong.model;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 
 public abstract class GameObject {
-    protected int x;
-    protected int y;
+    protected float x;
+    protected float y;
 
-    public GameObject(int x, int y) {
+    public GameObject(float x, float y) {
         this.x = x;
         this.y = y;
     }
